Hack for disabling rename channel and manage member links

+++ api/channel.go
@@ -437,38 +437,40 @@ func leave(c *Context, w http.ResponseWriter, r *http.Request) {
 }
 
 func deleteChannel(c *Context, w http.ResponseWriter, r *http.Request) {
+	l4g.Debug(utils.T("deleteChannel called: " + strconv.FormatBool(c.isSystemAdmin)))
+	if c.isSystemAdmin {
+		params := mux.Vars(r)
+		id := params["channel_id"]
 
-	params := mux.Vars(r)
-	id := params["channel_id"]
+		var channel *model.Channel
+		var err *model.AppError
+		if channel, err = app.GetChannel(id); err != nil {
+			c.Err = err
+			return
+		}
 
-	var channel *model.Channel
-	var err *model.AppError
-	if channel, err = app.GetChannel(id); err != nil {
-		c.Err = err
-		return
+		if channel.Type == model.CHANNEL_OPEN && !app.SessionHasPermissionToChannel(c.Session, channel.Id, model.PERMISSION_DELETE_PUBLIC_CHANNEL) {
+			c.SetPermissionError(model.PERMISSION_DELETE_PUBLIC_CHANNEL)
+			return
+		}
+
+		if channel.Type == model.CHANNEL_PRIVATE && !app.SessionHasPermissionToChannel(c.Session, channel.Id, model.PERMISSION_DELETE_PRIVATE_CHANNEL) {
+			c.SetPermissionError(model.PERMISSION_DELETE_PRIVATE_CHANNEL)
+			return
+		}
+
+		err = app.DeleteChannel(channel, c.Session.UserId)
+		if err != nil {
+			c.Err = err
+			return
+		}
+
+		c.LogAudit("name=" + channel.Name)
+
+		result := make(map[string]string)
+		result["id"] = channel.Id
+		w.Write([]byte(model.MapToJson(result)))
 	}
-
-	if channel.Type == model.CHANNEL_OPEN && !app.SessionHasPermissionToChannel(c.Session, channel.Id, model.PERMISSION_DELETE_PUBLIC_CHANNEL) {
-		c.SetPermissionError(model.PERMISSION_DELETE_PUBLIC_CHANNEL)
-		return
-	}
-
-	if channel.Type == model.CHANNEL_PRIVATE && !app.SessionHasPermissionToChannel(c.Session, channel.Id, model.PERMISSION_DELETE_PRIVATE_CHANNEL) {
-		c.SetPermissionError(model.PERMISSION_DELETE_PRIVATE_CHANNEL)
-		return
-	}
-
-	err = app.DeleteChannel(channel, c.Session.UserId)
-	if err != nil {
-		c.Err = err
-		return
-	}
-
-	c.LogAudit("name=" + channel.Name)
-
-	result := make(map[string]string)
-	result["id"] = channel.Id
-	w.Write([]byte(model.MapToJson(result)))
 }
 
 func getChannel(c *Context, w http.ResponseWriter, r *http.Request) {
